Add "Flag" labels/context to objective BIN values
Previous context labels were technically correct, but too ambiguous.

"Flag Captures" and "Flag Carrier Kills" represent the grand total amounts of objective captures (CTF/COL) and objective carrier kills (CTF/HRO/COL). The "Flag" part of it seems to be outdated.

It might've made sense early on in Warhawk's life. But when update 1.4 introduced the Hero/Collection modes, maybe "Flag" was repurposed to "Objective" in the SVML versions of those kinds of stats, only for the BIN labels to be overlooked.